# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Armi PDF/Word/Excel/PPT/TXT/RTF 檔案管理器 RESTful API。",
        "title": "Armi File Manager API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/files": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Retrieves a paginated list of file metadata records accessible to the authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "List user files",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by tag",
                        "name": "tag",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (starts from 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items per page (1-100)",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.FileListResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid pagination parameters",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Uploads a new file (PDF, Word, Excel, PPT, TXT, RTF) with upload progress reporting, SHA3-256 deduplication check, and embedding generation.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Upload a document file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "The document file to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.FileResponse"
                        }
                    },
                    "400": {
                        "description": "Missing file or invalid request",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "File conflict (duplicate filename or hash)",
                        "schema": {
                            "$ref": "#/definitions/contract.FileConflictResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/files/search": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Generates an embedding query and performs vector search for similar documents, returning metadata, similarity scores and source queries.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Search files semantically",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query text",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Max search results limit (default 5)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Enable NLP semantic search query expansion",
                        "name": "nlp_expansion",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of alternative queries to generate (default 3)",
                        "name": "expansion_num",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.SearchListResponse"
                        }
                    },
                    "400": {
                        "description": "Missing query parameter 'q'",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Search failed",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/mcp": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Handles MCP Streamable HTTP requests over a single /mcp endpoint.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mcp"
                ],
                "summary": "MCP Server (Streamable HTTP)",
                "responses": {
                    "200": {
                        "description": "MCP response",
                        "schema": {
                            "$ref": "#/definitions/http.StreamResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Handles MCP Streamable HTTP requests over a single /mcp endpoint.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mcp"
                ],
                "summary": "MCP Server (Streamable HTTP)",
                "responses": {
                    "200": {
                        "description": "MCP response",
                        "schema": {
                            "$ref": "#/definitions/http.StreamResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Handles MCP Streamable HTTP requests over a single /mcp endpoint.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mcp"
                ],
                "summary": "MCP Server (Streamable HTTP)",
                "responses": {
                    "200": {
                        "description": "MCP response",
                        "schema": {
                            "$ref": "#/definitions/http.StreamResponse"
                        }
                    }
                }
            }
        },
        "/users/me": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get current user profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.MeResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Register a new user in the system using Argon2id for password hashing.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "User Registration Details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contract.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.RegisterResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input or registration error",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Update current user profile",
                "parameters": [
                    {
                        "description": "Current user profile updates",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contract.UpdateMeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.MeResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Update failed",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "contract.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "contract.FileConflictResponse": {
            "type": "object",
            "properties": {
                "conflicting_hash": {
                    "type": "string"
                },
                "conflicting_id": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "contract.FileListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/contract.FileResponse"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "contract.FileResponse": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "string"
                },
                "content_type": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "embedding_status": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "group_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "hash": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "contract.MeResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "contract.RegisterRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "contract.RegisterResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "contract.SearchListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/contract.SearchResponseItem"
                    }
                },
                "nlp_expansion": {
                    "type": "boolean"
                }
            }
        },
        "contract.SearchResponseItem": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "string"
                },
                "chunk_id": {
                    "type": "string"
                },
                "chunk_text": {
                    "type": "string"
                },
                "content_type": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "distance": {
                    "type": "number"
                },
                "embedding_status": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "group_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "hash": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "score": {
                    "type": "number"
                },
                "size": {
                    "type": "integer"
                },
                "source_query": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "contract.UpdateMeRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "http.StreamResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}